requirements.txt
Flask
pyjwt
pycryptodome
colorama